x/sbt/types/message_create_soul_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"github.com/tbruyelle/nebula/testutil/sample"
)

func TestMsgCreateSoul_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateSoul
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateSoul{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateSoul{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
